Conflicts:
	GMU-Projekt/include/matrix_operations.h
	GMU-Projekt/src/main.cpp
	GMU-Projekt/src/matrix_operations.cl
	GMU-Projekt/src/matrix_operations.cpp

+++ GMU-Projekt/include/matrix_operations.h
@@ -1,3 +1,3 @@
-void determinant(float *, long double *, int *, int);
+void determinant(float *, float *, int *, int);
 void gem(float *, float *, int);
 void inverse(float *, float *, int);
